requirements.txt
# Core
fastapi==0.111.0
uvicorn==0.29.0
requests==2.32.4
python-dotenv==1.1.1

# Testing
pytest==8.4.1
pytest-asyncio==1.1.0

## Dashboard
streamlit==1.46.1
